Made a few important, unmentioned, discoveries.
(1)
ThetaX, the threshold used to determine if a neighbor is firing, needs to be significantly less than, ThetaY, the threshold used by the global inhibitor to determine if any neighbors are firing.

Excluding this detail ensures that a follower trying to respond to a leaders spike, will recieve a net stimulus of zero because it will get 1 from the leader at the same time it gets -1 from the global inhibitor. Adding a delay between ThetaX and ThetaY allows the follower to respond to the leader's stimulus before being inhibited.

(2)
The theta used as the threshold for lateral potential, to decide when to cut off the oscillator's input stimulus must be very small. Because it decays with the same epsilon rate as the inhibition, if this theta is any significant value (I was using .25) than the lateral potential will cancel out the input value for all oscilators before any can have a second spike. Making this smaller (I changed to .1) allows the oscilators, that are stimulated by both neighbors, time to wait for their inhibition to lower.

Trying to implement this paper has been extremely difficult. I am now beginning to realise that these differential equations do not actually work how they are written. And that these PHDs don't actually implement them. They use the Differential equations as a form of psuedo code that they then implement algorithmically.