Found the paper that explains how this is supposed to work. My problems with noise were because I was assuming a zero mean gaussian distribution, but the paper explained this should be -.02 to prevent the oscillators from self firing as I was experiencing.
This paper confirmed many of the things I figured out on my own, including the previous comment about theta thresholds, many of the variables I came to are plainly written is this much more useful paper.

This paper also explains that you don't actually implement the differential equations, except for the excitement of each oscillator, instead you algorithmically process the steps produced by the differential equations. For example, instead of updating all of the equations in time steps of .01, allowing oscilators to fire and impact their neighbors, you manually select which oscillator should fire next, find all of its neighbors that will fire in response, and continue steping through the process manually.

So in the end these huge frustrations ended up being just a learning experience, and now I see how it should actually be done. I was naive enough to assume we simulated every oscillator mathematically as discribed in the paper.

Since I've put so much effort into getting this "working" (I have 5 nodes FINNALY responding how they are supposed to), I would like to keep what I had done, even though it is extrememly slow, but it won't be practical to do it this way because it limits the maximum number of segments possible because there is literally not enough room on the waves produced by these differential equations.

So, back to the beginning I go, with a few days before demo.... but I learned a lot.